requirements.txt
nicegui
python-dotenv
supabase
httpx